refactor types and utils to fix circ dep
