+++ clients/fides-js/src/components/Overlay.tsx
@@ -1,6 +1,10 @@
 import { h, FunctionComponent, VNode } from "preact";
 import { useEffect, useState, useCallback, useMemo } from "preact/hooks";
-import { FidesOptions, PrivacyExperience } from "../lib/consent-types";
+import {
+  FidesCookie,
+  FidesOptions,
+  PrivacyExperience,
+} from "../lib/consent-types";
 
 import { debugLog, shouldResurfaceConsent } from "../lib/consent-utils";
 
@@ -9,7 +13,6 @@ import { useA11yDialog } from "../lib/a11y-dialog";
 import ConsentModal from "./ConsentModal";
 import { useHasMounted } from "../lib/hooks";
 import { dispatchFidesEvent } from "../lib/events";
-import { FidesCookie } from "../lib/cookie";
 import ConsentContent from "./ConsentContent";
 
 interface RenderBannerProps {
