+++ clients/fides-js/src/services/api.ts
@@ -3,6 +3,7 @@ import {
   ConsentMethod,
   EmptyExperience,
   FidesApiOptions,
+  FidesCookie,
   FidesOptions,
   PrivacyExperience,
   PrivacyPreferencesRequest,
@@ -10,7 +11,6 @@ import {
   RecordsServedResponse,
 } from "../lib/consent-types";
 import { debugLog } from "../lib/consent-utils";
-import { FidesCookie } from "../lib/cookie";
 
 export enum FidesEndpointPaths {
   PRIVACY_EXPERIENCE = "/privacy-experience",
